First pass at SceneFileWriter refactor
