+++ active_projects/clacks/all_questions_scenes.py
@@ -1,0 +1,8 @@
+from active_projects import clacks
+
+output_directory = "clacks_question"
+all_scenes = [
+    clacks.NameIntro,
+    clacks.MathAndPhysicsConspiring,
+    clacks.LightBouncing,
+]
