+++ manimlib/config.py
@@ -33,7 +33,7 @@ def parse_cli():
             help="Render the scene as a movie file",
         ),
         parser.add_argument(
-            "-s", "--show_last_frame",
+            "-s", "--save_last_frame",
             action="store_true",
             help="Save the last frame and open the image file",
         ),
@@ -73,7 +73,7 @@ def parse_cli():
             help="Write all the scenes from a file",
         ),
         parser.add_argument(
-            "-o", "--output_file_name",
+            "-o", "--file_name",
             help="Specify the name of the output file, if"
                  "it should be different from the scene class name",
         )
@@ -155,36 +155,25 @@ def get_module(file_name):
 
 
 def get_configuration(args):
-    if args.output_file_name is not None:
-        output_file_name_root, output_file_name_ext = os.path.splitext(
-            args.output_file_name)
-        expected_ext = '.png' if args.show_last_frame else '.mp4'
-        if output_file_name_ext not in ['', expected_ext]:
-            print("WARNING: The output will be to (doubly-dotted) %s%s" %
-                  output_file_name_root % expected_ext)
-            output_file_name = args.output_file_name
-        else:
-            # If anyone wants .mp4.mp4 and is surprised to only get .mp4, or such... Well, too bad.
-            output_file_name = output_file_name_root
-    else:
-        output_file_name = args.output_file_name
-
+    file_writer_config = {
+        # By default, write to file
+        "write_to_movie": args.write_to_movie or not args.save_last_frame,
+        "save_last_frame": args.save_last_frame,
+        "save_pngs": args.save_pngs,
+        # If -t is passed in (for transparent), this will be RGBA
+        "png_mode": "RGBA" if args.transparent else "RGB",
+        "movie_file_extension": ".mov" if args.transparent else ".mp4",
+        "file_name": args.file_name,
+    }
     config = {
         "module": get_module(args.file),
         "scene_names": args.scene_names,
         "open_video_upon_completion": args.preview,
         "show_file_in_finder": args.show_file_in_finder,
-        # By default, write to file
-        "write_to_movie": args.write_to_movie or not args.show_last_frame,
-        "show_last_frame": args.show_last_frame,
-        "save_pngs": args.save_pngs,
-        # If -t is passed in (for transparent), this will be RGBA
-        "saved_image_mode": "RGBA" if args.transparent else "RGB",
-        "movie_file_extension": ".mov" if args.transparent else ".mp4",
+        "file_writer_config": file_writer_config,
         "quiet": args.quiet or args.write_all,
         "ignore_waits": args.preview,
         "write_all": args.write_all,
-        "output_file_name": output_file_name,
         "start_at_animation_number": args.start_at_animation_number,
         "end_at_animation_number": None,
         "sound": args.sound,
@@ -241,7 +230,7 @@ def get_configuration(args):
             config["start_at_animation_number"] = int(stan)
 
     config["skip_animations"] = any([
-        config["show_last_frame"] and not config["write_to_movie"],
+        file_writer_config["save_last_frame"],
         config["start_at_animation_number"],
     ])
     return config
