+++ manimlib/scene/scene.py
@@ -1,16 +1,9 @@
-from time import sleep
-import _thread as thread
-import datetime
 import inspect
-import os
 import random
-import shutil
-import subprocess
 import warnings
 
 from tqdm import tqdm as ProgressDisplay
 import numpy as np
-from pydub import AudioSegment
 
 from manimlib.animation.animation import Animation
 from manimlib.animation.creation import Write
@@ -21,12 +14,8 @@ from manimlib.container.container import Container
 from manimlib.continual_animation.continual_animation import ContinualAnimation
 from manimlib.mobject.mobject import Mobject
 from manimlib.mobject.svg.tex_mobject import TextMobject
+from manimlib.scene.scene_file_writer import SceneFileWriter
 from manimlib.utils.iterables import list_update
-from manimlib.utils.output_directory_getters import add_extension_if_not_present
-from manimlib.utils.output_directory_getters import get_image_output_directory
-from manimlib.utils.output_directory_getters import get_movie_output_directory
-from manimlib.utils.output_directory_getters import get_partial_movie_output_directory
-from manimlib.utils.output_directory_getters import get_sorted_integer_files
 
 
 class Scene(Container):
@@ -34,20 +23,12 @@ class Scene(Container):
         "camera_class": Camera,
         "camera_config": {},
         "frame_duration": LOW_QUALITY_FRAME_DURATION,
-        "construct_args": [],
+        "file_writer_config": {},
         "skip_animations": False,
-        "write_to_movie": False,
-        "save_pngs": False,
-        "pngs_mode": "RGBA",
-        "movie_file_extension": ".mp4",
         "always_continually_update": False,
         "random_seed": 0,
         "start_at_animation_number": None,
         "end_at_animation_number": None,
-        "livestreaming": False,
-        "to_twitch": False,
-        "twitch_key": None,
-        "output_file_name": None,
         "leave_progress_bars": False,
     }
 
@@ -55,43 +36,36 @@ class Scene(Container):
         # Perhaps allow passing in a non-empty *mobjects parameter?
         Container.__init__(self, **kwargs)
         self.camera = self.camera_class(**self.camera_config)
+        self.file_writer = SceneFileWriter(
+            self, **self.file_writer_config,
+        )
+
         self.mobjects = []
         self.continual_animations = []
         self.foreground_mobjects = []
         self.num_plays = 0
-        self.frame_num = 0
         self.time = 0
         self.original_skipping_status = self.skip_animations
-        self.stream_lock = False
         if self.random_seed is not None:
             random.seed(self.random_seed)
             np.random.seed(self.random_seed)
 
-        self.init_audio()
         self.setup()
-        if self.livestreaming:
-            return None
         try:
-            self.construct(*self.construct_args)
+            self.construct()
         except EndSceneEarlyException:
-            if hasattr(self, "writing_process"):
-                self.writing_process.terminate()
+            pass
         self.tear_down()
-
-        if self.write_to_movie:
-            self.combine_movie_files()
+        self.file_writer.finish()
         self.print_end_message()
 
     def handle_play_like_call(func):
         def wrapper(self, *args, **kwargs):
             self.handle_animation_skipping()
-            should_write = self.write_to_movie and not self.skip_animations
-            if should_write:
-                self.open_movie_pipe()
-                func(self, *args, **kwargs)
-                self.close_movie_pipe()
-            else:
-                func(self, *args, **kwargs)
+            allow_write = not self.skip_animations
+            self.file_writer.begin_animation(allow_write)
+            func(self, *args, **kwargs)
+            self.file_writer.end_animation(allow_write)
             self.num_plays += 1
         return wrapper
 
@@ -116,11 +90,6 @@ class Scene(Container):
     def __str__(self):
         return self.__class__.__name__
 
-    def get_output_file_name(self):
-        if self.output_file_name is not None:
-            return self.output_file_name
-        return str(self)
-
     def print_end_message(self):
         print("Played {} animations".format(self.num_plays))
 
@@ -142,40 +111,9 @@ class Scene(Container):
     def get_attrs(self, *keys):
         return [getattr(self, key) for key in keys]
 
-    # Sound
-    def init_audio(self):
-        self.includes_sound = False
-
-    def create_audio_segment(self):
-        self.audio_segment = AudioSegment.silent()
-
-    def add_audio_segment(self, new_segment, time_offset=0):
-        if not self.includes_sound:
-            self.includes_sound = True
-            self.create_audio_segment()
-        segment = self.audio_segment
-        overly_time = self.get_time() + time_offset
-        if overly_time < 0:
-            raise Exception("Adding sound at timestamp < 0")
-
-        curr_end = segment.duration_seconds
-        new_end = overly_time + new_segment.duration_seconds
-        diff = new_end - curr_end
-        if diff > 0:
-            segment = segment.append(
-                AudioSegment.silent(int(np.ceil(diff * 1000))),
-                crossfade=0,
-            )
-        self.audio_segment = segment.overlay(
-            new_segment, position=int(1000 * overly_time)
-        )
-
-    def add_sound(self, sound_file, time_offset=0):
-        new_segment = AudioSegment.from_file(sound_file)
-        self.add_audio_segment(new_segment, 0)
+    # TODO, Scene file writer now handles sound
 
     # Only these methods should touch the camera
-
     def set_camera(self, camera):
         self.camera = camera
 
@@ -202,9 +140,9 @@ class Scene(Container):
             mobjects=None,
             background=None,
             include_submobjects=True,
-            dont_update_when_skipping=True,
+            ignore_skipping=True,
             **kwargs):
-        if self.skip_animations and dont_update_when_skipping:
+        if self.skip_animations and not ignore_skipping:
             return
         if mobjects is None:
             mobjects = list_update(
@@ -522,8 +460,6 @@ class Scene(Container):
 
     @handle_play_like_call
     def play(self, *args, **kwargs):
-        if self.livestreaming:
-            self.stream_lock = False
         if len(args) == 0:
             warnings.warn("Called Scene.play with no animations")
             return
@@ -558,22 +494,11 @@ class Scene(Container):
         else:
             self.continual_update(0)
 
-        if self.livestreaming:
-            self.stream_lock = True
-            thread.start_new_thread(self.idle_stream, ())
         return self
 
+    # TODO
     def idle_stream(self):
-        while(self.stream_lock):
-            a = datetime.datetime.now()
-            self.update_frame()
-            n_frames = 1
-            frame = self.get_frame()
-            self.add_frames(*[frame] * n_frames)
-            b = datetime.datetime.now()
-            time_diff = (b - a).total_seconds()
-            if time_diff < self.frame_duration:
-                sleep(self.frame_duration - time_diff)
+        self.file_writer.idle_stream()
 
     def clean_up_animations(self, *animations):
         for animation in animations:
@@ -638,201 +563,15 @@ class Scene(Container):
         return self
 
     def add_frames(self, *frames):
+        self.increment_time(len(frames) * self.frame_duration)
         if self.skip_animations:
             return
-        self.increment_time(len(frames) * self.frame_duration)
-        if self.write_to_movie:
-            for frame in frames:
-                if self.save_pngs:
-                    self.save_image(
-                        "frame" + str(self.frame_num), self.pngs_mode, True
-                    )
-                    self.frame_num = self.frame_num + 1
-                self.writing_process.stdin.write(frame.tostring())
-
-    # Display methods
+        for frame in frames:
+            self.file_writer.write_frame(frame)
 
     def show_frame(self):
-        self.update_frame(dont_update_when_skipping=False)
+        self.update_frame(ignore_skipping=True)
         self.get_image().show()
-
-    def get_image_file_path(self, name=None, dont_update=False):
-        sub_dir = "images"
-        output_file_name = self.get_output_file_name()
-        if dont_update:
-            sub_dir = output_file_name
-        path = get_image_output_directory(self.__class__, sub_dir)
-        file_name = add_extension_if_not_present(
-            name or output_file_name, ".png"
-        )
-        return os.path.join(path, file_name)
-
-    def save_image(self, name=None, mode="RGB", dont_update=False):
-        path = self.get_image_file_path(name, dont_update)
-        if not dont_update:
-            self.update_frame(dont_update_when_skipping=False)
-        image = self.get_image()
-        image = image.convert(mode)
-        image.save(path)
-
-    def get_movie_file_path(self, name=None, extension=None):
-        directory = get_movie_output_directory(
-            self.__class__, self.camera_config, self.frame_duration
-        )
-        if extension is None:
-            extension = self.movie_file_extension
-        if name is None:
-            name = self.get_output_file_name()
-        file_path = os.path.join(directory, name)
-        if not file_path.endswith(extension):
-            file_path += extension
-        return file_path
-
-    def get_partial_movie_directory(self):
-        return get_partial_movie_output_directory(
-            self, self.camera_config, self.frame_duration
-        )
-
-    def open_movie_pipe(self):
-        directory = self.get_partial_movie_directory()
-        file_path = os.path.join(
-            directory, "{}{}".format(
-                self.num_plays,
-                self.movie_file_extension,
-            )
-        )
-        temp_file_path = file_path.replace(".", "_temp.")
-
-        self.movie_file_path = file_path
-        self.temp_movie_file_path = temp_file_path
-
-        fps = int(1 / self.frame_duration)
-        height = self.camera.get_pixel_height()
-        width = self.camera.get_pixel_width()
-
-        command = [
-            FFMPEG_BIN,
-            '-y',  # overwrite output file if it exists
-            '-f', 'rawvideo',
-            '-s', '%dx%d' % (width, height),  # size of one frame
-            '-pix_fmt', 'rgba',
-            '-r', str(fps),  # frames per second
-            '-i', '-',  # The imput comes from a pipe
-            '-c:v', 'h264_nvenc',
-            '-an',  # Tells FFMPEG not to expect any audio
-            '-loglevel', 'error',
-        ]
-        if self.movie_file_extension == ".mov":
-            # This is if the background of the exported video
-            # should be transparent.
-            command += [
-                '-vcodec', 'qtrle',
-                # '-vcodec', 'png',
-            ]
-        else:
-            command += [
-                '-vcodec', 'libx264',
-                '-pix_fmt', 'yuv420p',
-            ]
-        if self.livestreaming:
-            if self.to_twitch:
-                command += ['-f', 'flv']
-                command += ['rtmp://live.twitch.tv/app/' + self.twitch_key]
-            else:
-                command += ['-f', 'mpegts']
-                command += [STREAMING_PROTOCOL + '://' + STREAMING_IP + ':' + STREAMING_PORT]
-        else:
-            command += [temp_file_path]
-        self.writing_process = subprocess.Popen(command, stdin=subprocess.PIPE)
-
-    def close_movie_pipe(self):
-        self.writing_process.stdin.close()
-        self.writing_process.wait()
-        if self.livestreaming:
-            return True
-        shutil.move(
-            self.temp_movie_file_path,
-            self.movie_file_path,
-        )
-
-    def combine_movie_files(self):
-        # Manim renders the scene as many smaller movie files
-        # which are then concatenated to a larger one.  The reason
-        # for this is that sometimes video-editing is made easier when
-        # one works with the broken up scene, which effectively has
-        # cuts at all the places you might want.  But for viewing
-        # the scene as a whole, one of course wants to see it as a
-        # single piece.
-        partial_movie_file_directory = self.get_partial_movie_directory()
-        kwargs = {
-            "remove_non_integer_files": True,
-            "extension": self.movie_file_extension,
-        }
-        if self.start_at_animation_number is not None:
-            kwargs["min_index"] = self.start_at_animation_number
-        if self.end_at_animation_number is not None:
-            kwargs["max_index"] = self.end_at_animation_number
-        else:
-            kwargs["remove_indices_greater_than"] = self.num_plays - 1
-        partial_movie_files = get_sorted_integer_files(
-            partial_movie_file_directory,
-            **kwargs
-        )
-        # Write a file partial_file_list.txt containing all
-        # partial movie files
-        file_list = os.path.join(
-            partial_movie_file_directory,
-            "partial_movie_file_list.txt"
-        )
-        with open(file_list, 'w') as fp:
-            for pf_path in partial_movie_files:
-                if os.name == 'nt':
-                    pf_path = pf_path.replace('\\', '/')
-                fp.write("file \'{}\'\n".format(pf_path))
-
-        movie_file_path = self.get_movie_file_path()
-        commands = [
-            FFMPEG_BIN,
-            '-y',  # overwrite output file if it exists
-            '-f', 'concat',
-            '-safe', '0',
-            '-i', file_list,
-            '-c', 'copy',
-            '-loglevel', 'error',
-            movie_file_path
-        ]
-        if not self.includes_sound:
-            commands.insert(-1, '-an')
-
-        combine_process = subprocess.Popen(commands)
-        combine_process.wait()
-        # os.remove(file_list)
-
-        if self.includes_sound:
-            sound_file_path = movie_file_path.replace(
-                self.movie_file_extension, ".wav"
-            )
-            # Makes sure sound file length will match video file
-            self.add_audio_segment(AudioSegment.silent(0))
-            self.audio_segment.export(sound_file_path)
-            temp_file_path = movie_file_path.replace(".", "_temp.")
-            commands = commands = [
-                "ffmpeg",
-                "-i", movie_file_path,
-                "-i", sound_file_path,
-                '-y',  # overwrite output file if it exists
-                "-c:v", "copy", "-c:a", "aac",
-                '-loglevel', 'error',
-                "-shortest",
-                "-strict", "experimental",
-                temp_file_path,
-            ]
-            subprocess.call(commands)
-            shutil.move(temp_file_path, movie_file_path)
-            # subprocess.call(["rm", self.temp_movie_file_path])
-            subprocess.call(["rm", sound_file_path])
-
-        print("\nAnimation ready at {}\n".format(movie_file_path))
 
     # TODO, this doesn't belong in Scene, but should be
     # part of some more specialized subclass optimized
